update manifest.json with version tag
Home Assistant will start requiring custom components to use version tags in the manifest json. Updating this manually with every pull request seems to be the easiest way to go about it since this custom component won't be updated that frequently. I used CalVer versioning with Full Year, Short Month, and Version as the template. Home Assistant requires that the version be a valid version recognized by AwesomeVersion, which CalVer is.