Add error handling and aborting to config_flow
Rewritten to include testing user provided credentials for successful log in ability. If unable to log in with user provided credentials, user is notified on the integration screen. Also includes check to make sure only one instance of the integration is installed at a time. Attempting to add a second instance of IOCare will abort the attempt and show an appropriate message to the end user.