Add Preset Modes
This code change adds auto mode and night mode as preset modes which can be called with "fan.set_preset_mode". Users are still able to call auto mode and night mode through a service call. 

Other changes: Calling a percentage > 0 when the purifier is off will now turn the purifier on and set it to the specified percentage. Previously, setting a percentage when the purifier was off did nothing. This also goes for auto mode and night mode. If the purifier is off when the auto mode and night mode presets are called, the purifier will turn on and be set to the specified preset mode.